Code wrote as good practic more  before
