+++ java-classes-practical/src/am/javaprogram/classes/studentlist/services/StudentService.java
@@ -1,43 +1,16 @@
 package am.javaprogram.classes.studentlist.services;
 
+import am.javaprogram.classes.studentlist.StudentGenerator;
 import am.javaprogram.classes.studentlist.models.Student;
-import am.javaprogram.classes.studentlist.interfaces.StudentGenerator;
 
-import java.util.Random;
-
-public class StudentService implements StudentGenerator {
+public class StudentService {
   private Student[] studentData = new Student[10];
 
   public StudentService() {
-    StudentGenerator generator;
-    String[] names = {
-      "Tigran", "Vardan", "Ashot", "Vahram", "Valod", "Gurgen", "Ishxan", "Varazdat", "Gevorg"
-    };
-    String[] faculties = {"IT", "ECONOMIC", "ELECTRONIC"};
-    generator =
-        () -> {
-          Student generate =
-              new Student(
-                  names[getRangeByMinMax(0, names.length - 1)],
-                  getRangeByMinMax(1994, 2000) + "-02-13",
-                  faculties[getRangeByMinMax(0, faculties.length - 1)],
-                  getRangeByMinMax(1, 4),
-                  getRangeByMinMax(800, 810));
-          return generate;
-        };
+    StudentGenerator generator = new StudentGenerator();
     for (int i = 0; i < studentData.length; i++) {
       studentData[i] = generator.genarateStudent();
     }
-  }
-
-  @Override
-  public Student genarateStudent() {
-    return null;
-  }
-
-  private int getRangeByMinMax(int min, int max) {
-    Random r = new Random();
-    return r.nextInt((max - min) + 1) + min;
   }
 
   public static void printFilteredData(Student[] array) {
@@ -83,7 +56,7 @@ public class StudentService implements StudentGenerator {
 
   public Student[] getByOldFromTargetYear(int year) {
     int countFoundData = 0;
-    Student[] foundData= new Student[studentData.length];
+    Student[] foundData = new Student[studentData.length];
     for (Student student : studentData) {
       if (student.getBirth().getYear() < year) {
         foundData[countFoundData++] = student;
